PD#147269: misc: backport per-UID statistics driver
misc: uidstat: change release handler for stat read operation

A kmalloced object in the file opening path is not kfreed in the
closing path.

In the struct file_operations, the open operation is assigned to
single_open while release is assigned to seq_release. single_open
internally allocates memory dynamically for an object  of type
seq_operations which is subsequently not freed in the release call
to seq_release.

Fix this by calling single_release which correctly frees this
memory allocated in the open call.

Change-Id: I3ea677be301244638e3843834c46988ad999ed88
Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>
(cherry picked from commit 85bee652d4a19e2441bdffc34294d1e797900784)
Signed-off-by: Yixun Lan <yixun.lan@amlogic.com>