PD#142598: security: fix security issue CVE-2014-9922
fs: limit filesystem stacking depth

Add a simple read-only counter to super_block that indicates how deep this
is in the stack of filesystems.  Previously ecryptfs was the only stackable
filesystem and it explicitly disallowed multiple layers of itself.

Overlayfs, however, can be stacked recursively and also may be stacked
on top of ecryptfs or vice versa.

To limit the kernel stack usage we must limit the depth of the
filesystem stack.  Initially the limit is set to 2.

Change-Id: I321cd0d4b198ff6e6a62e0b373b36c330e6f140c
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>