PD#147269: misc: backport per-UID statistics driver
uid_cputime: Avoids double accounting of process stime, utime and cpu_power in
task exit.

This avoids the race where a particular process is terminating and we
read the show_uid_stats. At this time since the task_struct still exists
and we will account for the terminating process as one of the active
task, where as the stats would have been added in the task exit
callback.

Bug: 22064385

Change-Id: If4c035b4c8350001f87d725fec344fd61b9e0688
Signed-off-by: Jin Qian <jinqian@google.com>
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>
Signed-off-by: Yixun Lan <yixun.lan@amlogic.com>