PD#147269: misc: backport per-UID statistics driver
uid_cputime: Check for the range while removing range of UIDs.

Checking if the uid_entry->uid matches the uid intended to be removed will
prevent deleting unwanted uid_entry.
Type cast the key for the hashtable to the same size, as when they were
inserted. This will make sure that we can find the uid_entry we want.

Bug: 25195548

Change-Id: Ic9c11a8c2f4d302c1b10801a2b7a7a20103258c5
Signed-off: Ruchi kandoi <kandoiruchi@google.com>
Signed-off-by: Yixun Lan <yixun.lan@amlogic.com>
Signed-off-by: Jianxin Pan <jianxin.pan@amlogic.com>