PD#115517: ion_carveout_heap: remove ion heap pages zero function
The reserved memory region for ion involve the logo region
set in uboot. The ion memory will be set to zero when carveout
heap created which will make screen black untill frame buffer
opened. Remove ion_heap_pages_zero() to work-around this issue.
This is the tempory solution of course.

Change-Id: If8192867f7081f107654211f72f1ac4f9fe7fd81
Signed-off-by: Zhenxing Jin <zhenxing.jin@amlogic.com>