PD#136759: ANDROID: dm: Add android verity target
cherry-pick from android-3.18: f559c8d3aa

This device-mapper target is virtually a VERITY target. This
target is setup by reading the metadata contents piggybacked
to the actual data blocks in the block device. The signature
of the metadata contents are verified against the key included
in the system keyring. Upon success, the underlying verity
target is setup.

BUG: 27175947

Change-Id: I33a6e306b22de7bb1b4c96f7b07769eed33aee05
Signed-off-by: Xindong Xu <xindong.xu@amlogic.com>