PD#131406: usb: rework xhci.
1. Clean the CEC flag. This flag indicates a change has occurred in the
 port’s Current Connect Status (CCS)or Cold Attach Status (CAS) bits.
Note that this flag shall not be set if the CCS transition was due to
software setting PP to ‘0’, or the CAS transition was due to software
setting WPR to ‘1’.The xHC sets this bit to ‘1’ for all changes to the
port device connect status, even if system software has not cleared an
existing Connect Status Change. For example, the insertion status
changes twice before system software has cleared the changed condition,
root hub hardware will be “setting” an already-set bit (i.e., the bit
will remain ‘1’). Software shall clear this bit by writing a ‘1’ to it.

2.If the xHCI host controller has died (ie, device removed) or suffered
other serious fatal error (STS_FATAL), then sofeware should remove xhci-hcd
driver and rework it.

Change-Id: I91ece09a2901b71aa3a8474a744c84fe8407620f