UPSTREAM: PM / devfreq: Fix available_governor sysfs
The devfreq using passive governor is not able to change the governor.
So, the user can not change the governor through 'available_governor' sysfs
entry. Also, the devfreq which don't use the passive governor is not able to
change to 'passive' governor on the fly.

Fixes: 996133119f57 ("PM / devfreq: Add new passive governor")
Cc: stable@vger.kernel.org
Signed-off-by: Chanwoo Choi <cw00.choi@samsung.com>
Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>
(cherry picked from commit bcf23c79c4e46130701370af4383b61a3cba755c)

Conflicts:
	drivers/devfreq/governor_passive.c

Change-Id: I70b888c018eb624f3ef5b500381ad1d227712352
Signed-off-by: Finley Xiao <finley.xiao@rock-chips.com>