amlogic: remote: Add support for Sony SIRC 12/15/20
Contrary to most IR RC protocols, the Sony SIRC frames end with a data
bit, not with a trailer pulse, which triggers the following bugs in the
hardware IR decoder because it relies on pulse intervals instead of
pulse widths:

 - The last data bit (msb) cannot be read properly. It can only be
   dropped, in which case it is always returned as 0, or forcibly read
   by using an abnormally long maximum frame duration and ignoring the
   logic 1 timings, in which case it is stuck to 1. This implementation
   chooses to drop the msb because it is part of the device ID, so this
   limitation is very unlikely to cause any trouble.

 - The repeat maximum frame interval in AO_MF_IR_DEC_REG3 is ignored.
   Consequently, if the same RC key is pressed twice in a row, the
   decoder reports the second key press as a repeat event, even if there
   is a very long time between the key presses. This implementation
   works around this bug by considering reported repeat events as actual
   normal events if they occur more than a given time after the previous
   event.

In order to support the SIRC protocol without any limitation, a software
decoder would have to be used with the general time measurement mode. To
this end, an equivalent of
drivers/amlogic/input/remote/remote_decoder_xmp.c for SIRC could be
derived from drivers/media/rc/ir-sony-decoder.c.

Signed-off-by: Benoît Thébaudeau <benoit@wsystem.com>