gpio: pca953x: Add support for Amlogic GPIO IRQs
Amlogic's GPIO driver handles interrupts in a non-standard way.
Therefore, all the drivers needing Amlogic GPIO IRQs have to be modified
in order to support them.

Here, the pca953x GPIO expander driver is modified to make it possible
to have its IRQ output wired to an Amlogic GPIO. The DTS files needing
this have to define the "irq-gpios" property as the Amlogic GPIO to
which the expander IRQ output is wired, e.g.:

&pinmux {
	gpio_expander0_pins: gpio-expander0-pins {
		amlogic,setmask=<6 0x00000000>;
		amlogic,clrmask=<6 0x00000000>;
		amlogic,pins = "GPIOH_5";
		amlogic,pullupen = <1>;
		amlogic,pullup = <1>;
		amlogic,enable-output = <0>;
	};
};

&i2c_b {
	gpio_expander0: gpio-expander@20 {
		compatible = "nxp,pca9555";
		reg = <0x20>;
		irq-gpios = <&gpio GPIOH_5 GPIO_ACTIVE_LOW>;
		interrupt-parent = <&gic>;
		interrupts = <GIC_SPI 64 IRQ_TYPE_LEVEL_HIGH>;
		pinctrl-names = "default";
		pinctrl-0 = <&gpio_expander0_pins>;
		gpio-controller;
		#gpio-cells = <2>;
		interrupt-controller;
		#interrupt-cells = <2>;
	};
};

Signed-off-by: Benoît Thébaudeau <benoit@wsystem.com>