input: rotary-encoder: Add support for Amlogic GPIO IRQs
Amlogic's GPIO driver handles interrupts in a non-standard way.
Therefore, all the drivers needing Amlogic GPIO IRQs have to be modified
in order to support them.

Here, the rotary encoder driver is modified to make it possible to have
its outputs wired to Amlogic GPIOs. The DTS files needing this have to
define the "interrupts" property as the list of the interrupts of the
Amlogic GPIOs to which the encoder outputs are wired, e.g.:

/ {
	rotary-encoder0 {
		compatible = "rotary-encoder";
		gpios = <&gpio_ao GPIOAO_6 GPIO_ACTIVE_HIGH>,
			<&gpio_ao GPIOAO_3 GPIO_ACTIVE_HIGH>;
		interrupts = <GIC_SPI 65 IRQ_TYPE_EDGE_RISING>,
			     <GIC_SPI 66 IRQ_TYPE_EDGE_RISING>,
			     <GIC_SPI 70 IRQ_TYPE_EDGE_RISING>,
			     <GIC_SPI 71 IRQ_TYPE_EDGE_RISING>;
		linux,axis = <7>; /* REL_DIAL */
		rotary-encoder,relative-axis;
		pinctrl-names = "default";
		pinctrl-0 = <&rotary_encoder0_pins>;
	};
};

&pinmux {
	rotary_encoder0_pins: rotary-encoder0-pins {
		amlogic,setmask=<AO 0x00000000>;
		amlogic,clrmask=<AO 0x00450280>;
		amlogic,pins = "GPIOAO_6", "GPIOAO_3";
		amlogic,pullupen = <0>;
		amlogic,enable-output = <0>;
	};
};

Signed-off-by: Benoît Thébaudeau <benoit@wsystem.com>