iio: adc: meson-saradc: enable new interface to return processed value [4/6]
PD#SWPL-19042

Problem:
the existing interface return only the raw sampling value which is related to
resolution, it is not convenient to use directly the raw value for consumer
driver because range of value is different.

Solution:
enable new interface to return processed value, no matter what the resolution
is, it is always return the 10-bit value.

Verify:
test pass on u200

Change-Id: Id5fec9b1050a09f5bcc35677c98caff480019ac9
Signed-off-by: Xingyu Chen <xingyu.chen@amlogic.com>