dts: time: rm arm,no-tick-in-suspend property [1/1]
PD#SWPL-18765

Problem:
Wall time is incorrect after suspend/resume without rtc

Solution:
Rely on Always on system counter

Verify:
AC200

arm,no-tick-in-suspend is a property that indicates the system counter
cannot keep counting after system suspend. However the systerm counter
for ARMv8 resides in Always Power On domain and is capable of remaining
counting even suspend. So the arm,no-tick-in-suspend must not be set

The arm,no-tick-in-suspend is added from
a55022706068c461f13552fbbb4e0f8cf056933f by khilman@baylibre.com, and
from the commit message "The Arm per-CPU architected timers stop ticking
in suspend, when the SCP powers down the CPUs.  Flag that in the DT." we
know khilman misunderstood the usage of arm,no-tick-in-suspend, aka,
arm,no-tick-in-suspend is for systerm counter not for per-CPU timer

Change-Id: Ide64b592e66cbeedcdde5e20ff5b48874bdd3dd1
Signed-off-by: Jiamin Ma <jiamin.ma@amlogic.com>