pwm: meson: Adds times, blink, and constant special functions [1/1]
PD#SWPL-19109

Problem:
In addition to the general pwm function, the amlogic pwm controller has
developed the times, blink, and constant functions based on the double-
channel function. Since these functions are unique to amlogic, they are
not defined in the pwm framework of linux and belong to non-standard
interfaces. These interfaces are declared in linux/amlogic/pwm-meson.h
and must be included when used.

Solution:
Adds times, blink, and constant special functions

Verify:
test pass on sm1_ac200_v1

Change-Id: Iac583ad126e2b961609cf3d843e8ce994db252f2
Signed-off-by: Bichao Zheng <bichao.zheng@amlogic.com>