pwm: fix frequency and duty not accurate in high frequencies [1/1]
PD#SWPL-19109

Problem:
The high and low settings of the duty register in the PWM
controller have been incremented by 1 on the controller.
When the pwm frequency is closer to the clock source,
high+low will be small, which will seriously affect the
accuracy of pwm.

Solution:
High and low Decrement one when setting the duty register

Verify:
test pass on sm1_ac213_v1

Change-Id: Idda27fb3e7c18668626283fa65378aa3879a83d7
Signed-off-by: Bichao Zheng <bichao.zheng@amlogic.com>