clk: the CCF use the glitch-free mali mux [1/1]
PD#SWPL-21424

Problem:
only use on one mux for gpu

Solution:
The "mali_0" or "mali_1" clock trees should not be updated while the
clock is running. Enforce this by setting CLK_SET_RATE_GATE on the
"mali_0" and "mali_1" gates. This makes the CCF switch to the "mali_1"
tree when "mali_0" is currently active and vice versa, which is exactly
what the vendor driver does when updating the frequency of the mali
clock.

Verify:
ac234#020

Change-Id: I60a6cda1a47efce8b6be24bb4b7c07b0bb92b44e
Signed-off-by: Jiyu Yang <Jiyu.Yang@amlogic.com>