remote: power key event is repeatedly reported when the system wakes up [1/1]
PD#SWPL-30120

Problem:
If we choose the legacy IR controller to respone the IR-KEY after system
standby, an invalid interrupt was triggered when the system wakes up, and
the interrupt will cause the power key event is repeatedly reported at the
wakeup phase.

Solution:
If the data frame is invalid, the interrupt handle should be ignored.

Verify:
test pass on txlx

Change-Id: Ia3359f0dd8ac18101d1327394d9b50c2eefe76f1
Signed-off-by: Xingyu Chen <xingyu.chen@amlogic.com>