audio: Atmos_indication_ddp-MP4 and Atmos_prompt_ddp-MP4 [2/2]
PD#SWPL-54218

Problem:
Atmos_indication_ddp-MP4 and Atmos_prompt_ddp-MP4

Solution:
When DAP is enable and dap_surround_virtualizer is OFF, the Dolby Atmos
indicator is not displayed. "-dap_surround_virtualizer 0,96" is disable
Dolby Atmos, when the Dolby-Atmos is coming, the TV prompts the user to
switch on Dolby Atmos for the best experience.

Verify:
on Askey's SBR, these two cases are passed.

Change-Id: Iadd66a814b5a3751aa0e506c965268fee96aec4f
Signed-off-by: Xiushan Lu <xiushan.lu@amlogic.com>