audio: fix audio driver stuck issue [1/1]
PD#SWPL-55853

Problem:
1. TDM same source open and write
2. SPDIF then open and write
As the tdm samesrc is still there but spdif output
is not ready to prepare(buffer level less than start threshold),
then the tdm output HW is waiting for spdif to start, but the spdif
writing thread(same writing thread to tdm out) is waiting for available
buffer to write. This comes to deadlock.

Solution:
Add release samesrc when spdif open if the tdm same source is on.

Verify:
SC2

Change-Id: I146a94e15b3c8cd738715ef0fe2fd308639b5a82
Signed-off-by: Shuai Li <shuai.li@amlogic.com>