mm: pin locked file after fault [1/1]
PD#SWPL-57667

Problem:
If we pin a file in android, the whole data of this file
will be loaded to DDR, but most of these data may not be
used, this caused memory waste.

Solution:
delay mark mlocked flags for unevictable pages until
it has been faulted.
You can use following command to disable this funciton:
    echo 0 > /proc/sys/vm/shrink_unevictable

Verify:
t5d am311

Signed-off-by: Tao Zeng <tao.zeng@amlogic.com>
Change-Id: I373f50ce6712f00d421846ca07672bb7cb3029ca