hdmitx21: adjust the VPU_HDMI_SETTING setting [2/2]
PD#SWPL-60141

Problem:
The decouple-FIFO between VENCI/P and hdmi may be in disorder when y420
mode switching

Solution:
When video frequency change, firstly clear the enable bits to 0,
and after frequency are settled, set the enable bit to 1.
The decouple-FIFO between VENCI/P and hdmi will be reset and put
the read pointer of the FIFO to a new safe position.

Verify:
t7/an400

Change-Id: Iea089452ce77474675037fe46f46e870ec4156d5
Signed-off-by: Zongdong Jiao <zongdong.jiao@amlogic.com>