drm: rework SET_MASTER and DROP_MASTER perm handling [1/1]
PD#SWPL-65409

Problem:
the SET/DROP_MASTER ioctls used to return EACCES when CAP_SYS_ADMIN was not
set

Solution:
This commit reworks the permission handling of the two ioctls. In
particular it enforced the CAP_SYS_ADMIN check only, if:
 - we're issuing the ioctl from process other than the one which opened
the node, and
 - we are, or were master in the past

This ensures that we:
 - do not regress the systemd-logind style of DRM_MASTER arbitrator
 - allow applications which do not use systemd-logind to drop their
master capabilities (and regain them at later point) ... w/o running as root.

Verify:
S4_AP222

Change-Id: I41387bfe65b40bf5724bef46b5197ae9615d36e5
Signed-off-by: Bing Jiang <bing.jiang@amlogic.com>