sched: rt: select rt task elegant [1/1]
PD#SWPL-86765

Problem:
By default wakeup a rt task will try to select the cpu which the task
last run on, it's friendly to high prio rt task performance, but it may
directly preempt high-prio CFS task or lower-prio rt task which is
running on the cpu currently, so it's not kind to these slightly lower-prio
important tasks.

Solution:
select rt task more elegant, if current task prio higher than a certain value
(eg: 106), then try to find a cpu which the currently running task prio is
lowest on it.

Verify:
s4

Change-Id: I92094019c7e90a94ec2cf2ab89ddfc59d8928b4b
Signed-off-by: Hanjie Lin <hanjie.lin@amlogic.com>