codec_mm: CB2 Scatter memory leakage. [1/1]
PD#SWPL-86524

Problem:
Probability of scatter memory leak
when switching resolutions frequently.
The reason is that the scatter map is an array.
Since the buffer is not limited,
the number of buffers exceeds the upper limit of the array.
When alloc scatter memory,
there is no map array record.
And the release is unsuccessful, memory leaks occur.

Solution:
Record the applied scatter context with linked list.

Verify:
sc2

Change-Id: Ibd070d117cd45fdff5841fdd0e145d44ed9b4b59
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>