codec_mm: CB2 Scatter memory leakage. [1/1]
PD#SWPL-86524

Problem:
Probability of scatter memory leak
when switching resolutions frequently.
The reason is that the scatter map is an array.
Since the buffer is not limited,
the number of buffers exceeds the upper limit of the array.
When alloc scatter memory,
there is no map array record.
And the release is unsuccessful, memory leaks occur.

Solution:
increase map array size.

Verify:
sc2

Change-Id: I9dda94462f52e1607fa8da33b2165c2a492e4fe4
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>