ANDROID: arm64: Reclaim some cpucaps to aid in backporting
Although we padded the cpucaps space for android13-5.10, the move to a
generated header in android13-5.15 was missed and consequently we are
unable to backport some errata workarounds from upstream without
breaking the KMI.

To resolve this problem, reclaim some cpucaps allocated by errata
workarounds which will never be relevant to Android (i.e. those
which apply only to server parts such as Cavium designs and the
Neoverse-N1) and which have their use carefully confined to core
architecture code.

Bug: 248633761
Signed-off-by: Will Deacon <willdeacon@google.com>
Change-Id: Id2d9429cae62d97fee44d051318aa5e5b93b29d8