amdv: color abnormal when switch from simply south to netflix [1/1]
PD#SWPL-96413

Problem:
Video decoder instance from simplysouth is not released, then the
second video instance from netflix is created.But the second inst
is not mapped in DV module and reused the first dv inst.

Solution:
Map the second instance for the second video. The second inst is
primary video and display on vd1, and only this instance is valid
for DV module when only one dv core1 is supported.

Test:
local test ok

Verify:
sc2

Signed-off-by: yao liu <yao.liu@amlogic.com>
Change-Id: I941743ed9504191fb93edbeb97593036a7cffd0b