eth: wol: fix wake up if wol feature is disabled
Now it set every time the wol wake up MAC address and enable the PMEB pull if
wol feature is enabled. This will wake up the device if the feature is disabled
as the bootloader blob trigger on gpio low level. Change it to enable PMEB pull
up by default on boot and do set MAC only if wol feature is enabled.

Signed-off-by: Ivan Li <ivan.li@wesion.com>
Change-Id: I411f0087824d67530a365b8f7f2ced069ec8150b