mm: pin locked file after fault. [1/2]
PD#SWPL-99264

Problem:
If we pin a file in android, the whole data of this file
will be loaded to DDR, but most of these data may not be
used, this caused memory waste.

Solution:
when pinning a file, do not load the file,
only wait until a page fault occurs later to load it.

Verify:
s4_aq222.

Change-Id: Ife507989e0233332579bd8df072ac105d9683daf
Signed-off-by: Jianxiong Pan <jianxiong.pan@amlogic.com>