Update kvims_defconfig
nf_tables is the default firewall framework on Ubuntu 22.04 and Debian 11 (and others too). This commit adds the main module and others related to it.

According to my tests, you can't even use iptables-legacy or the other tools to manage the firewall.

I've built an image with the Fenix script for the VIM1 including those modules and all the firewall commands and tools work as expected in Ubuntu 22.04