ASoC: hdmi-codec: Stop stream when plug state changed
Workaround for HDMIIN and HDMIOUT plug-{in,out} when streaming.

Actually, we should do stop stream both for HDMI_{OUT,IN} on
plug-{out,in} event. but for better experience and depop stream,
we optimize as follows:

a) Do stop stream for HDMIIN on plug-out when streaming.
because HDMIIN work as SLAVE mode, CLK lost after HDMI cable
plugged out which will make stream stuck until ALSA timeout(10s).
so, for better experience, we should stop stream at the moment.

b) Do stop stream for HDMIOUT on plug-in when streaming.
because HDMIOUT work as MASTER mode, there is no clk-issue like
HDMIIN, but, on HDR situation, HDMI will be reconfigured which
make HDMI audio configure lost, especially for NLPCM/HBR bitstream
which require IEC937 packet alignment, so, for this situation,
we stop stream to notify user to re-open and configure sound card
and then go on streaming.

Userspace should notice Error-Code(such as -EBADFD) from snd_pcm_write
to do reopen.

Signed-off-by: Sugar Zhang <sugar.zhang@rock-chips.com>
Change-Id: I85455e56f843afcdc9a5f1d5c9e85733b6732cdf