net: rfkill-bt: add power up func when waked by bt-irq only.
when rtk(also some others) wifi/bt combo chips suspend, it
disconnect from bt-remote control unit(RCU). when RCU's
powerkey pressed, it will send ble broadcast to rtk bt chip.

As the rtk chips receive the broadcast, it should wake up
the host, but it can't send bt-hid powerkey to host because
it had lost communication with host.
So,it just send a IRQ pulse to host. The host should resume
from suspend.

Because the RCU just send a IRQ pulse, so rk chips just resume
by IRQ and then go to sleep again for no power-key event.
this is not expected. we expect to power up system wholly.

With this path, we try to power up the system. so we add a
power-up key event when BT IRQ received.

We also can control the "/proc/bluetooth/sleep/powerupkey"
node to enable or disable this function.

Change-Id: Ie59b4a2c4cd2f91820d31835df86565003126465
Signed-off-by: Zou Dengming <marsow.zou@rock-chips.com>