v4l: CF2 Adjust gfp_flag according to the order. [1/1]
PD#SWPL-131346

Problem:
When the memory is in short supply, if the order is too large,
the memory will not be allocated at one time,
which will increase the allocation time.

Solution:
According to the degree of memory shortage,
adjust the size of the order, so as to allocate memory
in small amounts and multiple times to improve efficiency.

Verify:
t5m

Change-Id: I2d6900d115285dc606841428f349a0a588fa67fa
Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>