media: CF1 add feature for dmabuf status dumping. [1/1]
PD#SWPL-137630

Problem:
codec mm adds the interface for dmabuf status dumping.

Solution:
The dmabuf infos interface is mainly used to view
the status of dmabuf and track the status that dmabuf
is referenced by a process. It can provide a convenient
way for memory leakage and find the leaking process.

Verify:
ohm

Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>
Change-Id: I0ab8a121b26badc2f8037cad56bd69073db3189d