arm: set TEXT_OFF default with 0x208000 [1/1]
PD#SWPL-144091

Problem:
swapper_pg_dir is the virtual address of the initial page table.
It will use this value as the starting address for the reserved memory.
the value depend on TEXT_OFF in arm, and default is 0x8000.
In 5.4, CONFIG_ARCH_MESON be set, it be set 0x208000, but
in 5.15, CONFIG_ARCH_MESON not set, so the start addr conflict with dmc.

Solution:
set TEXT_OFF default with 0x208000

Verify:
planck

Change-Id: Iea66ca0edcd360f9f54e8df3574553170403d571
Signed-off-by: qinglin.li <qinglin.li@amlogic.com>