zram: fix crash when cpu off line [1/1]
PD#SWPL-145196

Problem:
kernel NULL pointer panic when cpu off line

Solution:
The root cause is that context for zram is configured as 2 part:
one for direct reclaim and one for kswapd. They are not for per-cpu
but when a core is off, zram will free per-cpu context so caused panic
on other cpus. So we should remove free code for crypto hooks.

Verify:
T5M

Change-Id: I65b9535d5f958dd272845f808b1d4099359cbfc7
Signed-off-by: Jianxiong Pan <jianxiong.pan@amlogic.com>