hdmitx: Adjust the disable phy timing during mode switching [1/1]
PD#SWPL-144353

Problem:
During the mode switching process, when systemctrl writes the
node to close phy, the phy action is to close it immediately.
At this time, part of the video data may have been transmitted,
but the data is incomplete. At this time, rx got part of the image.
Some TVs did not perform frame dropping processing after getting part
of the data and parsed it, but the solution was wrong, which may
cause some display abnormalities.

Solution:
Referring to the processing method of the Apple box,
during the mode switching process, when systemctrl issues
a shutdown phy command, we will wait until vsync is
interrupted before shutting down the phy. At this time,
there is no video data transmission, which is relatively safe.

Verify:
ohm/T7

Change-Id: I36bdc75271db7a6d1b7842b6b952ac657250c12a
Signed-off-by: xiang.wu1 <xiang.wu1@amlogic.com>