WIP: arm64: dts: meson: drop broadcom compatible from reference board SDIO nodes
Remove the Broadcom compatible to allow Android STB boards using Qualcom QCA9377
and Realtek RTL8189ES/FS SDIO modules to also have working WiFi when booting with
the reference board device-tree(s). Also do the same to the Vega S95 dtsi as this
is commonly used in the same way.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>