video: rockchip: mpp: rkvenc2: clean irq_status after read split len
In vepu_510 the sli_done interrupt is triggered by slice_fifo not empty
status.Therefore when the sli_done_sta interrupt is cleared without
reading all the slice_len data in slice_fifo an extra interrupt will
be triggered. So it is better to clear the interrupt after reading all
slice_len to avoid false irq reporting. On the other hand, vepu_580 is
triggered only after writing a slice, without checking iif fifo is
non-empty, so there is no need to clear the interrupt again after reading the slice

Change-Id: If20a097b0e9fb41df9984c0ed316d717ba783ac4
Signed-off-by: Yanjun Liao <yanjun.liao@rock-chips.com>