v4l: Adjust gfp_flag according to the order [1/1]
PD#SWPL-153208
PD#SWPL-131346

Problem:
1.When the memory is in short supply, if the
order is too large, the memory will not be
allocated at one time, which will increase
the allocation time.
2.Modify in vb2 did not following gki2.0

Solution:
1.According to the degree of memory shortage,
adjust the size of the order, so as to allocate
memory in small amounts and multiple times to
improve efficiency.
2.The modify in vb2 only for versions prior to
androidT and do not need to follow gki2.0. Now
use the new config to pass gki check.

Verify:
ohm

Change-Id: I4be66b8b06f9870041afefe9506d1e9f480c0bdd
Signed-off-by: Hao Shi <hao.shi@amlogic.com>