video: rockchip: mpp: rkvdec2: Set ip func to def val
Vdpu383 doesn't reset the register when it is enabled.
In order to avoid exceptions, it is necessary to
configure the initial value when it is enabled.

In addition, ip_en involves cru mode, which is related
to the IP work frequency. It is necessary to ensure
that it is configured correctly.

Signed-off-by: Hongjin Li <vic.hong@rock-chips.com>
Change-Id: I7047faababe1747547d18cb51e1e3b527e2c8d35