genirq: irq_pipeline: skip IRQ forwarding when running oob
A subsequent oob event might arrive before its handler had a chance to
process a pending forwarded interrupt from the in-band stage.

Since we don't want to delay interrupts when oob handlers exist, we
cannot mask them and/or defer EOI until the in-band stage did process
them. To address the issue, let's ignore forwarding if the oob stage
is current in forward_irq_event().

As a result, the semantics of IRQ forwarding requires the oob-side of
the interrupt handler to decode and record all forwarded events as
many times as it might be called in a row, until the in-band side is
able to process all of them in one go eventually.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>