drm/rockchip: vop2: get the correctly vp id when close all pd
The win attached vp maybe unused, and the unused vp will not
register crtc, so we need to lool up all win under the pd and check
if the win attached vp has register crtc.
If all the win attached vp is unused, we set the first win(undter
the pd) attach to the first crtc to make sure the pd can be closed normally.

Signed-off-by: Sandy Huang <hjc@rock-chips.com>
Change-Id: Id7dfedae2d3b6e7a139ca2eebaefea250e986d12