drm/rockchip: dw_hdmi: Don't set enable-gpio value when hdmi bind
If uboot logo is enabled and the enable-gpio value is
different from that in uboot, kernel logo will be displayed
abnormally. Therefore, we do not set the value of enable-gpio
when hdmi bind, but set it in encoder enable when hdmi is switched
after system boot. This method is also compatible with
scenarios that uboot logo is not enabled.

Change-Id: I238aeb63790b67414036e2c89382870ff575c4e5
Signed-off-by: Algea Cao <algea.cao@rock-chips.com>