drm/rockchip: dw_hdcp2: always support unlocked_ioctl() file operation
In both the Android SDK and Linux SDK, when compiling for a 64-bit
system, the CONFIG_COMPAT macro is enabled, but the unlocked_ioctl call
is not compiled. The hdcp2_tx_rx service compiled on Android is 32-bit,
corresponding to a 64-bit kernel. Therefore, ioctl calls go through
compat_ioctl without any issues. However, services and kernels built
with the NVR SDK and Linux SDK are consistent at 64 bits. This causes
ioctl calls to go through unlocked_ioctl, leading to exceptions.

Fixes: 8f36a88f84ff ("drm/rockchip: Add dw hdcp2 controller driver")
Change-Id: I3090bcb03e509360b54194d1152cf05437dc0627
Signed-off-by: Eddy Zhang <eddy.zhang@rock-chips.com>