arm64: dts: rockchip: Fixes warning in unit address of drm-logo for rk3576-linux.dtsi
/reserved-memory/drm-logo@00000000 refers to a node in the device tree.
The part after the @ symbol is the unit address, which should match the
physical or logical address of the device or memory region being described.

The warning indicates that the unit name (in this case, 00000000) should
not have leading zeros. This is more about adhering to conventions and
ensuring compatibility and readability than about functional correctness.

Change-Id: I85adad8caac6e2bfef0703640fecb6ff2c20b4f8
Signed-off-by: Caesar Wang <wxt@rock-chips.com>