ARM: dts: rockchip: rv1126: Fixes warning in unit address of drm-logo
/reserved-memory/drm-logo@00000000 refers to a node in the device tree.
The part after the @ symbol is the unit address, which should match the
physical or logical address of the device or memory region being
described.

The warning indicates that the unit name (in this case, 00000000) should
not have leading zeros. This is more about adhering to conventions and
ensuring compatibility and readability than about functional
correctness.

Signed-off-by: Sandy Huang <hjc@rock-chips.com>
Change-Id: I5052698913a2ff8202dbc0f8c2144acd48cb4ac5