drm/panel: panel-simple: delete pm_runtime
In upstream patch, runtime pm is used to avoid excessive unprepare/
prepare.

However, some DSI panels need to be powered off and on again before
reloading the initialization sequence. If set the auto_suspend and
delay time is set to 1s, the suspend/resume callback function cannot
be entered when sleeping/waking up quickly, the power of panel cannot
be restarted, causing this timing to be abnormal and causing the panel
to not work properly.

In kernel 5.10, pm_runtime has not been used in panel-simple, but it
still runs stably in panel applications. Considering the above
reasons, we chose to revert the submission of pm_runtime upstream.

Type: Fix
Redmine ID: 460825 468692
Associated modifications:
    commit 3235b0f20a0a ("drm/panel: panel-simple: Use runtime pm to
    avoid excessive unprepare / prepare")
Test: N/A

Signed-off-by: Zhibin Huang <zhibin.huang@rock-chips.com>
Change-Id: I348e1a860e528cf4bee7ce76de0ea13c18fddd90