drm/rockchip: logo: call .mode_fixup() before mode comparison
The mode comparison helps to confirm whether the display
mode of uboot logo is valid of the drm mode list in kernel.

The mode stored in dts has been fixed by .mode_fixup()
in uboot, which helps to deal with some workarounds, so
it is also needed to call .mode_fixup() when getting mode
list in kernel.

Change-Id: I2c74e66365a5e305ec8322316fdb7472e6df9fcd
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>