scsi: ufs: rockchip: Fix stability issues and add WB support
Disable UFS related clock in runtime is proved to be incorrect and
mismatch state between host and device should be recovered by reset the
controller to make sure each part is in the beginning status.

Meanwhile, we have verified WB works fine by test, so add WB support
as well.

Change-Id: I5af675576ef2b43ebe5061af204ba1fff5f3f233
Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>